binded forgotten editor specific buttons dungeon temple and portal
