+++ source/modes/EditorMode.cpp
@@ -37,6 +37,7 @@
 #include "render/RenderManager.h"
 #include "camera/CameraManager.h"
 #include "rooms/RoomManager.h"
+#include "rooms/RoomType.h"
 #include "sound/MusicPlayer.h"
 #include "network/ODClient.h"
 #include "network/ODServer.h"
@@ -135,6 +136,8 @@ EditorMode::EditorMode(ModeManager* modeManager):
             CEGUI::Event::Subscriber(&EditorMode::showQuitMenuFromOptions, this)
     ));
 
+    // Connect editor specific buttons (Rooms, traps, spells, tiles, lights, ...)
+
     //Map light
     connectGuiAction(Gui::EDITOR_MAPLIGHT_BUTTON,
                      AbstractApplicationMode::GuiAction::ButtonPressedMapLight);
@@ -146,6 +149,20 @@ EditorMode::EditorMode(ModeManager* modeManager):
     connectTileSelect(Gui::EDITOR_LAVA_BUTTON,TileVisual::lavaGround);
     connectTileSelect(Gui::EDITOR_ROCK_BUTTON,TileVisual::rockGround);
     connectTileSelect(Gui::EDITOR_WATER_BUTTON,TileVisual::waterGround);
+
+    addEventConnection(
+        mRootWindow->getChild(Gui::BUTTON_TEMPLE)->subscribeEvent(
+          CEGUI::PushButton::EventClicked,
+          CEGUI::Event::Subscriber(RoomSelector(RoomType::dungeonTemple, mPlayerSelection))
+        )
+    );
+
+    addEventConnection(
+        mRootWindow->getChild(Gui::BUTTON_PORTAL)->subscribeEvent(
+          CEGUI::PushButton::EventClicked,
+          CEGUI::Event::Subscriber(RoomSelector(RoomType::portal, mPlayerSelection))
+        )
+    );
 
     updateFlagColor();
